add a generic_rpc
