+++ djAerolith/base/views.py
@@ -33,7 +33,7 @@ from base.models import WordList, Lexicon, AlphagramTag
 from base.utils import (generate_question_map_from_alphagrams,
                         generate_question_list_from_alphagrams,
                         question_list_from_probabilities)
-from lib.macondo_interface import anagram_letters
+from lib.macondo_interface import anagram_letters, MacondoError
 from lib.response import response, StatusCode
 
 logger = logging.getLogger(__name__)
@@ -329,5 +329,8 @@ def listmanager(request):
 def word_lookup(request):
     lexicon = request.GET.get('lexicon')
     letters = request.GET.get('letters')
-    results = anagram_letters(lexicon, letters)
+    try:
+        results = anagram_letters(lexicon, letters)
+    except MacondoError as e:
+        return response(str(e), StatusCode.BAD_REQUEST)
     return response(results)
